+++ backend/src/api/mod.rs
@@ -7,3 +7,4 @@ mod config_watch;
 mod serve;
 pub(crate) mod hdhomerun_ssdp;
 pub(crate) mod hdhomerun_proprietary;
+mod sys_usage;