+++ frontend/src/app/components/dashboard/stats_view.rs
@@ -1,5 +1,6 @@
 use yew::prelude::*;
 use yew_i18n::use_translation;
+use shared::utils::human_readable_byte_size;
 use crate::app::components::{Card, PlaylistProgressStatusCard, StatusCard, StatusContext};
 
 #[function_component]
@@ -44,6 +45,13 @@ pub fn StatsView() -> Html {
         }
     };
 
+    let (mem, cpu) = status_ctx.system_info.as_ref().map_or_else(|| ("n/a".to_string(), "n/a".to_string()),
+        |system| (format!("{} / {}", human_readable_byte_size(system.memory_usage), human_readable_byte_size(system.memory_total)), format!("{:.2}%", system.cpu_usage)));
+
+
+    let (cache, users, connections) = status_ctx.status.as_ref().map_or_else(|| ("n/a".to_string(),"n/a".to_string(),"n/a".to_string()),
+         |status| (status.cache.as_ref().map_or_else(|| "n/a".to_string(), |c| c.clone()), status.active_users.to_string(), status.active_user_connections.to_string()));
+
     html! {
       <div class="tp__stats">
         <div class="tp__stats__header">
@@ -51,15 +59,16 @@ pub fn StatsView() -> Html {
         </div>
         <div class="tp__stats__body">
             <div class="tp__stats__body-group">
-                <Card><StatusCard title={translate.t("LABEL.MEMORY")} data={status_ctx.status.as_ref().map_or_else(|| "n/a".to_string(), |status| status.memory.clone())} /></Card>
-                <Card><StatusCard title={translate.t("LABEL.CACHE")} data={status_ctx.status.as_ref().map_or_else(|| "n/a".to_string(), |status| status.cache.as_ref().map_or_else(|| "n/a".to_string(), |c| c.clone()))} /></Card>
+                <Card><StatusCard title={translate.t("LABEL.MEMORY")} data={mem} /></Card>
+                <Card><StatusCard title={translate.t("LABEL.CPU")} data={cpu} /></Card>
+                <Card><StatusCard title={translate.t("LABEL.CACHE")} data={cache} /></Card>
             </div>
             <div class="tp__stats__body-group">
                 <Card><PlaylistProgressStatusCard /></Card>
             </div>
             <div class="tp__stats__body-group">
-                <Card><StatusCard title={translate.t("LABEL.ACTIVE_USERS")} data={status_ctx.status.as_ref().map_or_else(|| "n/a".to_string(), |status| status.active_users.to_string())} /></Card>
-                <Card><StatusCard title={translate.t("LABEL.ACTIVE_USER_CONNECTIONS")} data={status_ctx.status.as_ref().map_or_else(|| "n/a".to_string(), |status| status.active_user_connections.to_string())} /></Card>
+                <Card><StatusCard title={translate.t("LABEL.ACTIVE_USERS")} data={users} /></Card>
+                <Card><StatusCard title={translate.t("LABEL.ACTIVE_USER_CONNECTIONS")} data={connections} /></Card>
                 { render_active_provider_connections() }
             </div>
         </div>